Update to wikimedia/less.php 5.0
Fix Less_Parser::AllParsedFiles -> getParsedFiles().

This was previously calling the static method Less_Parser::AllParsedFiles, but
using a non-standard casing for the method name, and via the instance rather
than static access to the static function.

This has been deprecated in 4.x and removed in 5.0 in favour of getParsedFiles,
which does what the code previously expected already.